Development: Apply signal-API ESLint rule to all client code
The Angular client is fully migrated to signal-based APIs, so the per-module
allowlist is no longer needed. Rename enforce-signal-apis-in-migrated-modules ->
enforce-signal-apis and apply it to all code under src/main/webapp/app (including
co-located specs); legacy decorators (@Input/@Output/@ViewChild/...) are now
forbidden everywhere in the client.
